requirements.txt
Django==1.9.10
django-jet==1.0.2
djangorestframework==3.5.1
simplejson==3.9.0
#argparse==1.2.1
#wsgiref==0.1.2
